# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a time-of-use pricing service for EV chargers.",
        "title": "Presto TOU Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/chargers/bulk/schedules": {
            "post": {
                "description": "Update schedules for multiple chargers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chargers"
                ],
                "summary": "Bulk update schedules",
                "parameters": [
                    {
                        "description": "Bulk update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.BulkUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/chargers/{id}/price": {
            "get": {
                "description": "Get the price for a specific charger at a given timestamp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chargers"
                ],
                "summary": "Get price for a charger at a specific time",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Charger ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Timestamp in ISO8601 format (e.g., 2024-01-15T10:30:00Z)",
                        "name": "timestamp",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PriceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/chargers/{id}/schedules": {
            "get": {
                "description": "Get all schedules for a specific charger",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chargers"
                ],
                "summary": "Get schedules for a charger",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Charger ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.TOUSchedule"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Replace all schedules for a specific charger",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chargers"
                ],
                "summary": "Update all schedules for a charger",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Charger ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Array of schedules",
                        "name": "schedules",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.TOUSchedule"
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update a partial schedule for a specific charger",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chargers"
                ],
                "summary": "Patch a schedule for a charger",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Charger ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Schedule object",
                        "name": "schedule",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.TOUSchedule"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.BulkUpdateRequest": {
            "type": "object",
            "properties": {
                "charger_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "schedules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.TOUSchedule"
                    }
                }
            }
        },
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.PriceResponse": {
            "type": "object",
            "properties": {
                "charger_id": {
                    "type": "string"
                },
                "local_charger_time": {
                    "type": "string"
                },
                "price_per_kwh": {
                    "type": "number"
                },
                "pricing_period": {
                    "$ref": "#/definitions/domain.PricingPeriod"
                },
                "requested_time": {
                    "type": "string"
                }
            }
        },
        "domain.PricingPeriod": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "string"
                },
                "price_per_kwh": {
                    "type": "number"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
        "domain.TOUSchedule": {
            "type": "object",
            "properties": {
                "end_time": {
                    "description": "Format: \"HH:MM\"",
                    "type": "string"
                },
                "price_per_kwh": {
                    "type": "number"
                },
                "start_time": {
                    "description": "Format: \"HH:MM\"",
                    "type": "string"
                }
            }
        }
    }
}